resolves #2 add a task to get a slug from a blog post document title
