+++ package.json
@@ -11,7 +11,8 @@
     "publish-images": "node tasks/publish-images.js",
     "publish-blog-post": "node tasks/publish-blog-post.js",
     "init": "node tasks/init.js",
-    "init-tweets": "node tasks/init-tweets.js"
+    "init-tweets": "node tasks/init-tweets.js",
+    "get-slug": "node tasks/get-slug.js"
   },
   "repository": {
     "type": "git",
